All compiles in the mac.

+++ src/cmph_benchmark.c
@@ -111,8 +111,10 @@ void bm_end(const char* name) {
   timeval_subtract(&stime, &benchmark->end.ru_stime, &benchmark->begin.ru_stime);
   
   printf("Benchmark: %s\n", benchmark->name);
-  printf("User time used  : %ld.%06ld\n", utime.tv_sec, utime.tv_usec);
-  printf("System time used: %ld.%06ld\n", stime.tv_sec, stime.tv_usec);
+  printf("User time used  : %ld.%06ld\n",
+         utime.tv_sec, (long int)utime.tv_usec);
+  printf("System time used: %ld.%06ld\n",
+         stime.tv_sec, (long int)stime.tv_usec);
   printf("\n");
 }
